abstract shared tests between lists
