+++ src/components/calcite-pick-list/shared-list-tests.ts
@@ -1,4 +1,4 @@
-import { newE2EPage } from "@stencil/core/testing";
+import { E2EElement, E2EPage, newE2EPage } from "@stencil/core/testing";
 
 export const tests = {
   selectionAndDeselection(listType) {
@@ -48,7 +48,9 @@ export const tests = {
         </calcite-${listType}-list>`);
 
         const numSelected = await page.evaluate((listType) => {
-          const list: HTMLCalcitePickListElement = document.querySelector(`calcite-${listType}-list`);
+          const list: HTMLCalcitePickListElement | HTMLCalciteValueListElement = document.querySelector(
+            `calcite-${listType}-list`
+          );
           return list.getSelectedItems().then((result) => {
             return result.size;
           });
@@ -58,7 +60,7 @@ export const tests = {
       });
     });
     describe("shift click behavior", () => {
-      it.only("should multi-select", async () => {
+      it("should multi-select", async () => {
         const page = await newE2EPage();
         await page.setContent(`<calcite-${listType}-list multiple>
           <calcite-${listType}-list-item value="one" text-label="One"></calcite-${listType}-list-item>
@@ -74,10 +76,11 @@ export const tests = {
         await page.keyboard.down("Shift");
         await item3.click();
         await page.keyboard.up("Shift");
-        await page.waitFor(5000);
 
         const numSelected = await page.evaluate((listType) => {
-          const list: HTMLCalcitePickListElement = document.querySelector(`calcite-${listType}-list`);
+          const list: HTMLCalcitePickListElement | HTMLCalciteValueListElement = document.querySelector(
+            `calcite-${listType}-list`
+          );
           return list.getSelectedItems().then((result) => {
             return result.size;
           });
@@ -85,6 +88,141 @@ export const tests = {
 
         expect(numSelected).toBe(3);
       });
+    });
+  },
+  filterBehavior(listType) {
+    let page: E2EPage = null;
+    let item1: E2EElement;
+    let item2: E2EElement;
+    let item1Visible;
+    let item2Visible;
+    beforeEach(async () => {
+      page = await newE2EPage();
+      await page.setContent(`<calcite-${listType}-list filter-enabled="true">
+        <calcite-${listType}-list-item value="1" text-label="One" text-description="uno"></calcite-${listType}-list-item>
+        <calcite-${listType}-list-item value="2" text-label="Two" text-description="dos"></calcite-${listType}-list-item>
+      </calcite-${listType}-list>`);
+      item1 = await page.find(`calcite-${listType}-list-item[value="1"]`);
+      item2 = await page.find(`calcite-${listType}-list-item[value="2"]`);
+      item1.setProperty("metadata", { category: "first" });
+      item2.setProperty("metadata", { category: "second" });
+      await page.waitForChanges();
+      await page.evaluate((listType) => {
+        (window as any).filter = document
+          .querySelector(`calcite-${listType}-list`)
+          .shadowRoot.querySelector("calcite-filter");
+        const filter = (window as any).filter;
+        (window as any).filterInput = filter.shadowRoot.querySelector("input");
+      }, listType);
+    });
+    it("should match text in the text-label prop", async () => {
+      // Match first item
+      await page.evaluate(() => {
+        const filterInput = (window as any).filterInput;
+        filterInput.value = "one";
+        filterInput.dispatchEvent(new Event("input"));
+      });
+      await item2.waitForNotVisible();
+
+      item1Visible = await item1.isVisible();
+      item2Visible = await item2.isVisible();
+
+      expect(item1Visible).toBe(true);
+      expect(item2Visible).toBe(false);
+
+      // Match second item
+      await page.evaluate(() => {
+        const filterInput = (window as any).filterInput;
+        filterInput.value = "two";
+        filterInput.dispatchEvent(new Event("input"));
+      });
+      await item1.waitForNotVisible();
+
+      item1Visible = await item1.isVisible();
+      item2Visible = await item2.isVisible();
+      expect(item1Visible).toBe(false);
+      expect(item2Visible).toBe(true);
+    });
+    it("should match text in the text-description prop", async () => {
+      // Match first item
+      await page.evaluate(() => {
+        const filterInput = (window as any).filterInput;
+        filterInput.value = "uno";
+        filterInput.dispatchEvent(new Event("input"));
+      });
+      await item2.waitForNotVisible();
+
+      item1Visible = await item1.isVisible();
+      item2Visible = await item2.isVisible();
+
+      expect(item1Visible).toBe(true);
+      expect(item2Visible).toBe(false);
+
+      // Match second item
+      await page.evaluate(() => {
+        const filterInput = (window as any).filterInput;
+        filterInput.value = "dos";
+        filterInput.dispatchEvent(new Event("input"));
+      });
+      await item1.waitForNotVisible();
+
+      item1Visible = await item1.isVisible();
+      item2Visible = await item2.isVisible();
+      expect(item1Visible).toBe(false);
+      expect(item2Visible).toBe(true);
+    });
+    it("should match text in the metadata prop", async () => {
+      await page.evaluate(() => {
+        const filterInput = (window as any).filterInput;
+        filterInput.value = "first";
+        filterInput.dispatchEvent(new Event("input"));
+      });
+      await item2.waitForNotVisible();
+
+      let item1Visible = await item1.isVisible();
+      let item2Visible = await item2.isVisible();
+      expect(item1Visible).toBe(true);
+      expect(item2Visible).toBe(false);
+
+      await page.evaluate(() => {
+        const filterInput = (window as any).filterInput;
+        filterInput.value = "second";
+        filterInput.dispatchEvent(new Event("input"));
+      });
+      await item1.waitForNotVisible();
+
+      item1Visible = await item1.isVisible();
+      item2Visible = await item2.isVisible();
+      expect(item1Visible).toBe(false);
+      expect(item2Visible).toBe(true);
+    });
+  },
+  disabledStates(listType) {
+    it("disabled", async () => {
+      const page = await newE2EPage();
+      await page.setContent(`<calcite-${listType}-list disabled>
+        <calcite-${listType}-list-item value="one" text-label="One"></calcite-${listType}-list-item>
+      </calcite-${listType}-list>`);
+
+      const list = await page.find(`calcite-${listType}-list`);
+      const item1 = await list.find("[value=one]");
+      const toggleSpy = await list.spyOnEvent("calciteListChange");
+
+      await item1.click();
+      expect(toggleSpy).toHaveReceivedEventTimes(0);
+    });
+    it("loading", async () => {
+      const page = await newE2EPage();
+      await page.setContent(`<calcite-${listType}-list loading>
+        <calcite-${listType}-list-item value="one" text-label="One"></calcite-${listType}-list-item>
+      </calcite-${listType}-list>`);
+
+      const list = await page.find(`calcite-${listType}-list`);
+      const item1 = await list.find("[value=one]");
+      const toggleSpy = await list.spyOnEvent("calciteListChange");
+
+      await item1.click();
+      expect(toggleSpy).toHaveReceivedEventTimes(0);
     });
   }
 };
